Move status prop to new StatusEntity subclass
